% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitterRANN-package.R
\docType{package}
\name{jitterRANN-package}
\alias{jitterRANN-package}
\alias{jitterRANN}
\title{Wrapper for Arya and Mount's Approximate Nearest Neighbours (ANN) C++ library}
\description{
Finds the k nearest neighbours for every point in a given dataset in O(N log N) time using Arya and Mount's ANN library (v1.1.3). There is support for approximate as well as exact searches, fixed radius searches and bd as well as kd trees.
}
\seealso{
\code{\link{nn2}}
}
\keyword{package}
